feat: add SwapSlippage components for managing slippage settings
- Introduced `SlippageInput` and `SlippageSettingDialog` components to handle user input for slippage tolerance in swaps.
- Implemented validation logic for slippage input values to ensure they meet specified criteria.
- Updated the `useSwapSlippageActions` hook to utilize the new `SlippageSettingDialog` for improved user experience.
- Removed the deprecated `SwapSlippageContentContainer` component to streamline the codebase.
