+++ packages/components/src/SwapSlippage/index.ts
@@ -1,0 +1,2 @@
+export { default } from './SlippageSettingDialog';
+export { SlippageInput } from './SlippageInput';
