+++ packages/kit/src/views/Swap/hooks/useSwapSlippageActions.tsx
@@ -4,6 +4,7 @@ import { useIntl } from 'react-intl';
 
 import type { IDialogInstance } from '@onekeyhq/components';
 import { Dialog } from '@onekeyhq/components';
+import SlippageSettingDialog from '@onekeyhq/components/src/SwapSlippage';
 import { useSettingsAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
 import accountUtils from '@onekeyhq/shared/src/utils/accountUtils';
@@ -15,7 +16,6 @@ import {
 } from '@onekeyhq/shared/types/swap/types';
 
 import { useSwapSlippageDialogOpeningAtom } from '../../../states/jotai/contexts/swap';
-import SwapSlippageContentContainer from '../pages/components/SwapSlippageContentContainer';
 
 import { useSwapAddressInfo } from './useSwapAccount';
 import { useSwapSlippagePercentageModeInfo } from './useSwapState';
@@ -58,7 +58,7 @@ export function useSwapSlippageActions() {
     dialogRef.current = Dialog.show({
       title: intl.formatMessage({ id: ETranslations.slippage_tolerance_title }),
       renderContent: (
-        <SwapSlippageContentContainer
+        <SlippageSettingDialog
           swapSlippage={slippageItem}
           autoValue={autoValue}
           onSave={slippageOnSave}
