+++ packages/kit/src/views/Swap/pages/components/SwapHeaderRightActionContainer.tsx
@@ -24,6 +24,7 @@ import {
   HeaderButtonGroup,
   HeaderIconButton,
 } from '@onekeyhq/components/src/layouts/Navigation/Header';
+import { SlippageInput } from '@onekeyhq/components/src/SwapSlippage';
 import useAppNavigation from '@onekeyhq/kit/src/hooks/useAppNavigation';
 import { useSwapTypeSwitchAtom } from '@onekeyhq/kit/src/states/jotai/contexts/swap';
 import {
@@ -56,7 +57,6 @@ import { useSwapSlippagePercentageModeInfo } from '../../hooks/useSwapState';
 import { SwapProviderMirror } from '../SwapProviderMirror';
 
 import ProviderManageContainer from './ProviderManageContainer';
-import { SlippageInput } from './SwapSlippageContentContainer';
 
 const SwapSettingsCommonItem = ({
   value,
